+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/App.kt
@@ -2,6 +2,8 @@ package org.virtuslab.bazelsteward.app
 
 import mu.KotlinLogging
 import org.virtuslab.bazelsteward.core.AppConfig
+import org.virtuslab.bazelsteward.core.DependencyKind
+import org.virtuslab.bazelsteward.core.FileFinder
 import org.virtuslab.bazelsteward.core.GitHostClient
 import org.virtuslab.bazelsteward.core.GitHostClient.Companion.PrStatus.*
 import org.virtuslab.bazelsteward.core.common.GitOperations
