+++ app/src/main/kotlin/org/virtuslab/bazelsteward/app/MavenDependencyKind.kt
@@ -1,7 +1,8 @@
 package org.virtuslab.bazelsteward.app
 
 import mu.KotlinLogging
-import org.virtuslab.bazelsteward.core.library.Library
+import org.virtuslab.bazelsteward.core.DependencyKind
+import org.virtuslab.bazelsteward.core.PathPattern
 import org.virtuslab.bazelsteward.core.library.Version
 import org.virtuslab.bazelsteward.core.replacement.Heuristic
 import org.virtuslab.bazelsteward.core.replacement.VersionOnlyHeuristic
